+++ src/App.tsx
@@ -4,9 +4,9 @@ import { State } from "./state/core";
 import Landing from "./pages/Landing/Landing";
 import Overview from "./pages/Overview/Overview";
 import ProjectDetailsView from "./pages/ProjectDetailsView/ProjectDetailsView";
-import { PageHeader } from "./views/PageHeader";
+import { PageHeader } from "./composites/PageHeader";
 import { Grid, Box } from "@radix-ui/themes";
-import SideMenu from "./views/SideMenu";
+import SideMenu from "./composites/SideMenu";
 import ProjectComparisonChart from "./pages/ProjectComparisons/CompareProjects";
 import * as ScrollArea from "@radix-ui/react-scroll-area";
