+++ src/composites/ProjectManager/ProjectFileList.tsx
@@ -3,7 +3,7 @@ import { IconButton, Table, Flex, Text, Heading } from "@radix-ui/themes";
 import { Version } from "../../state";
 import { useState } from "react";
 import { PaginationButtons } from "../../pages/Overview/Overview";
-import SearchBar from "../../views/SearchBar";
+import SearchBar from "../SearchBar";
 import VersionFilters, { Filters } from "./Filters/Filters";
 import { matchSorter } from "match-sorter";
 import { DateRange } from "react-day-picker";
